admin/requirements.txt
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
Django==4.0.3
django-braces==1.15.0
django-cors-headers==3.11.0
django-js-asset==2.0.0
django-oauth-toolkit==1.7.0
django-rest-authemail==2.1.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
install==1.3.5
jsonschema==3.2.0
jwcrypto==1.0
oauthlib==3.2.0
paramiko==2.9.2
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.8
websocket-client==0.59.0
wrapt==1.14.0
